+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added experimental `--defer-constraints` option
 - Added experimental `--disable-user-triggers` option
 - Added experimental `--disable-integrity` option
+- Improved error message for older libpq
 
 ## 0.5.4 (2020-05-09)